refactor: better kernels + reorganization

+++ src/CPU/harris.hpp
@@ -1,3 +1,4 @@
+#include <cstddef>
 #include <tuple>
 #include <vector>
 
@@ -6,4 +7,4 @@ unsigned char** morphoErode(unsigned char** img, int width, int height);
 float **morphoDilate(float **img, int width, int height);
 float **computeHarrisResponse(unsigned char **img, int width, int height);
 unsigned char** harrisThreshold(float **harris, int width, int height, float threshold);
-std::vector<std::tuple<float, int, int>> detectHarrisPoints(unsigned char **image, int width, int height, int max_keypoints, float threshold);
+std::vector<std::tuple<float, int, int>> detectHarrisPoints(unsigned char **image, int width, int height, size_t max_keypoints, float threshold);
